+++ docs/workflow-rbac.md
@@ -27,7 +27,7 @@ rules:
       - patch
 ```
 
-For <= v3.3
+For <= v3.3 use.
 
 ```yaml
 apiVersion: rbac.authorization.k8s.io/v1
@@ -43,6 +43,8 @@ rules:
       - get
       - patch
 ```
+
+Warning: For many organisations, it may not be acceptable to give a workflow the `pod patch` permission, see [#3961](https://github.com/argoproj/argo-workflows/issues/3961)
 
 If you are not using the emissary, you'll need additional permissions.
 See [executor](https://github.com/argoproj/argo-workflows/tree/master/manifests/quick-start/base/executor) for suitable
